feat!: rename constants
This is for clarity, and to allow in future having both
NO_COMPRESSION_32 and NO_COMPRESSION_64, to allow clients to fully
specify that they do not want any _64 code data.

Note that LibreOffice doesn't support zip64 at the time of writing, so
this is still an issue. This is not just for "legacy" software.