feat: no need for local zip64 extra when streaming
This was added in
https://github.com/uktrade/stream-zip/commit/62f8cf1ee242f7d9fe1467fe9a93b20164764068,
but it might not really have ever been needed, so removing it saves a few
bytes. And stronger than that,
https://github.com/libarchive/libarchive/issues/1834 gives an argument that it
even makes the ZIP file invalid.

To give even more weight to the argument that this is ok, I made a few Zip64
files with data descriptors with InfoZIP. They did not have local Zip64 extra
fields. If this were a problem, it probably would have been discovered by now.