Fix overflow errors writing zip64 central dir
Previously, if all files are zip32, but the total size of the zip
exceeds 4GiB, stream-zip would throw an overflow error, because it would
try to write a zip32 central directory and/or end-of-central-directory
record. This would fail because those records contain offsets over 4GiB,
which cannot be represented in zip32 versions of those records.

This change upgrades to zip64 central directory and/or
end-of-central-directory records if the caller
provides the `allow_upgrade_to_64=True` kwarg.