docs: document how to make an async interface
This was discussed in https://github.com/uktrade/stream-zip/issues/87. Maybe
down the line this code will be brought into stream-zip, or an interface
created that doesn't use threads under the hood. However for now, documenting
how to work with what there is I think is the quickest/easiest step forward.