docs: reference to-file-like-obj package rather than repeating code
This removes the code from a recipe - converting the iterable output to a
file-like object, and replaces it with an import from
https://github.com/uktrade/to-file-like-obj

This should hopefully make it just a touch easier for people to use the
recipes. It's deliberate to not include the to-file-like-obj in the
dependencies of this project in order to keep the dependencies down - not
everyone will need it.