ci: bump version of pytest
Without this for some reason the following error can happen if when we add
other dependencies in later commits:

> ImportError: cannot import name 'Config' from 'pytest'

(Specifically the dependencies of pycryptodome and pyzipper)

Have to admit I'm not really sure why, and haven't spent that long trying to
work out why, but I think it's reasonable to use a more recent version even
without this issue and not really worth the time investigating it.

We're not using the most recent version of pytest because I don't think it can
be installed with Python 3.6.7 that we're still testing on. Will like to
maintain support for older Python for as long as possible to not cause
annoyances for people that are restricted to older Python but want the latest
stream-zip.