refactor: flags defined for each mode in single place
This reduces a little big of duplication for each mode - the flags are now
defined in a single line. Also, the flags are defined as a 2-byte unsigned
integer rather than a 2-byte bytes array.

This changes allow for the flags to be more easily dynamic for each mode - set
in a single place and then use binary or | to combine flags. This will be
useful in the upcoming chnage that adds AES encryption for each mode that will
require the first bit of the flags to be set if encryption is enabled.