docs: another link to warnings and better examples on password generation
This changes documentation in what is now the "Password protection /
encryption" section. It also changes tests to make sure the example in the
documentation works, and makes the tests themselves use better practices, in
case they are seen down the line as examples to follow in the real world.

Specifically, it links to https://crypto.stackexchange.com/a/109269/113464
which serves as a bit of a warning, but also contains some advice. And, it
makes the password in the example longer and random.

It does not implement all the recommendations at
https://crypto.stackexchange.com/a/109269/113464 at least for now to keep this
change small while I consider them - they may come in later changes (e.g.
renaming password to passphrase or generating a default ourselves).

This is done as part of the request/discussion at
https://github.com/uktrade/stream-zip/issues/93