build(deps): move to govuk-eleventy-plugin v6.0.3
This moves to the govuk-eleventy-plugin for documentation to the current
latest, v6.0.3

There are a few breaking changes, documented at
https://x-govuk.github.io/govuk-eleventy-plugin/upgrading/5-to-6/. The trikiest
of these is the change of how the organisation name is now inside the SVG logo.
This required a few iterations of:

- Editing the logo with Inkscape
- Manually editing the SVG / associated CSS
- Using an SVG minifier to reduce its size

The logo is now a bit smaller in size than it was even though it also has the
"DBT" text in it, but from what I can tell looking at it, no loss of quality.
So a small step forward on that front even ignoring the version bump.

I did spend some time trying to put the organisation name in as HTML text in
the page, but it required too much faff with CSS to get it looking right. It's
essentially fighting the latest version of GOV.UK frontend which expects it to
be part of the logo.