docs: simplify async interface
This simplifies the code a bit in the suggested async interface. This is
inspired by the discussion at https://github.com/uktrade/stream-unzip/pull/80,
and makes this code more consistent with the code there.

Whether this will continue to be an example in the docs or integrated into the
codebase in some way, I'm not sure right now.