feat: more type annotations
This still isn't a complete set - strict type annotation won't work, but
hopefully a step forward. The set of annotations is slightly arbitrary.

Inspired by the discussion at
https://github.com/uktrade/stream-zip/discussions/120

Meta: Deciding that this is more a 'feat' than a 'refactor' in terms of the
commit message. While it's not a runtime behaviour change, it is a de-facto
change of the API