feat: a Method has a clearer definition (that passes strict type checking)
Each "Method" now derives from a base class, that has a "get" function to get
the internal tuple that controls its file's behaviour. This makes it possible
for the same class to both have such a method, _and_ be callable to get a
different method, which is how the current NO_COMPRESSION_* types work as
methods but also as a callable to return a different method.

(Maybe this can be turned into a Protocol in future changes)