feat: type annotations now allow strict type checking
This changes the internal tuple type to make the uncompressed size and crc_32
non optional ints which makes it possible to pass type checking without any ifs
or casts. This is still slightly odd maybe because it means we hard code them
to 0 when they're not used. A "better" structure might be to have some sort of
more dynamic structure to avoid this - maybe multiple tuples or a more complex
class. However, suspect this is still a small step forward since it allows
strict type checking.