ci: bump checkout action to latest version
Mostly just since I noticed it was out of date, but also I think it does some
interactions with tokens and things, and makes it easier to use to use the
GitHub CLI in later step. Not sure if the earlier version does that, but we'll
need to bump at some point, so may as well be now.