ci: release process now includes a source but with current version
The source does not actually include the correct release version to avoid
having to change the source code based on the current or next version. While
this is maybe odd, it makes releasing quite easy. However, as
https://github.com/uktrade/stream-zip/issues/136 this makes it slightly harder
to release for Debian.

So we make a version of the source code that has the correct release version,
and add it to the GitHub release. This makes it so any tinkering with versions
is just done in this repository.